els-ASAP auto commit file/tag pcs-0.9.148-7.el6.src.rpm

+++ src/pcsd/ssl.rb
@@ -2,38 +2,15 @@ require 'rubygems'
 require 'webrick'
 require 'webrick/https'
 require 'openssl'
-require 'logger'
 require 'rack'
 
-def get_rhel_version()
-  if File.exists?('/etc/system-release')
-    release = File.open('/etc/system-release').read
-    match = /(\d+)\.(\d+)/.match(release)
-    if match
-      return match[1, 2].collect{ | x | x.to_i}
-    end
-  end
-  return nil
-end
+require 'bootstrap.rb'
+require 'pcs.rb'
 
-def is_rhel6()
-  version = get_rhel_version()
-  return (version and version[0] == 6)
-end
-
-def is_systemctl()
-  if File.exist?('/usr/bin/systemctl')
-    return true
-  else
-    return false
-  end
-end
-
-CRT_FILE = "/var/lib/pcsd/pcsd.crt"
-KEY_FILE = "/var/lib/pcsd/pcsd.key"
 server_name = WEBrick::Utils::getservername
+$logger = configure_logger('/var/log/pcsd/pcsd.log')
 
-if not File.exists?(CRT_FILE) or not File.exists?(KEY_FILE)
+def generate_cert_key_pair(server_name)
   name = "/C=US/ST=MN/L=Minneapolis/O=pcsd/OU=pcsd/CN=#{server_name}"
   ca   = OpenSSL::X509::Name.parse(name)
   key = OpenSSL::PKey::RSA.new(2048)
@@ -46,42 +23,65 @@ if not File.exists?(CRT_FILE) or not File.exists?(KEY_FILE)
   crt.not_before = Time.now
   crt.not_after  = Time.now + 10 * 365 * 24 * 60 * 60 # 10 year
   crt.sign(key, OpenSSL::Digest::SHA256.new)
+  return crt, key
+end
 
+if not File.exists?(CRT_FILE) or not File.exists?(KEY_FILE)
+  crt, key = generate_cert_key_pair(server_name)
   File.open(CRT_FILE, 'w',0700) {|f| f.write(crt)}
   File.open(KEY_FILE, 'w',0700) {|f| f.write(key)}
+else
+  crt, key = nil, nil
+  begin
+    crt = File.read(CRT_FILE)
+    key = File.read(KEY_FILE)
+  rescue => e
+    $logger.error "Unable to read certificate or key: #{e}"
+  end
+  crt_errors = verify_cert_key_pair(crt, key)
+  if crt_errors and not crt_errors.empty?
+    crt_errors.each { |err| $logger.error err }
+    $logger.error "Invalid certificate and/or key, using temporary ones"
+    crt, key = generate_cert_key_pair(server_name)
+  end
 end
 
 webrick_options = {
   :Port               => 2224,
-  :BindAddress        => nil,
+  :BindAddress        => '::',
+  :Host               => '::',
   :SSLEnable          => true,
   :SSLVerifyClient    => OpenSSL::SSL::VERIFY_NONE,
-  :SSLCertificate     => OpenSSL::X509::Certificate.new(File.open(CRT_FILE).read),
-  :SSLPrivateKey      => OpenSSL::PKey::RSA.new(File.open(KEY_FILE).read()),
+  :SSLCertificate     => OpenSSL::X509::Certificate.new(crt),
+  :SSLPrivateKey      => OpenSSL::PKey::RSA.new(key),
   :SSLCertName        => [[ "CN", server_name ]],
   :SSLOptions         => OpenSSL::SSL::OP_NO_SSLv2 | OpenSSL::SSL::OP_NO_SSLv3,
 }
 
-if is_systemctl
-  webrick_options[:StartCallback] = Proc.new {
-    `python /usr/lib/pcsd/systemd-notify-fix.py`
-  }
-end
-
 server = ::Rack::Handler::WEBrick
 trap(:INT) do
   puts "Shutting down (INT)"
-  server.shutdown
-  #exit
+  if server.instance_variable_get("@server")
+    server.shutdown
+  else
+    exit
+  end
 end
 
 trap(:TERM) do
   puts "Shutting down (TERM)"
-  server.shutdown
-  #exit
+  if server.instance_variable_get("@server")
+    server.shutdown
+  else
+    exit
+  end
 end
 
 require 'pcsd'
 begin
+  server.run(Sinatra::Application, webrick_options)
+rescue Errno::EAFNOSUPPORT
+  webrick_options[:BindAddress] = '0.0.0.0'
+  webrick_options[:Host] = '0.0.0.0'
   server.run(Sinatra::Application, webrick_options)
 end
